Minor changes caught during soft-fork in other repo.
Dedupe project name in Makefile.

Silence lint in main.c instead of casting bool to int for a bool
evaluation. Once the fix is available in the latest clang-tidy I can
remove the NOLINTNEXTLINE.

Clarify what the link is for in minor.yml.

+++ module_kernel/src/main.c
@@ -95,6 +95,6 @@ int module_stop(SceSize args, const void* argp) {
 
     LOG_INFO("Stopped");
 
-    // TODO drop (int) cast on new clang-tidy: https://github.com/llvm/llvm-project/issues/195604
-    return (int)failed ? SCE_KERNEL_STOP_FAIL : SCE_KERNEL_STOP_SUCCESS;
+    // NOLINTNEXTLINE(readability-implicit-bool-conversion) TODO remove https://github.com/llvm/llvm-project/issues/195604
+    return failed ? SCE_KERNEL_STOP_FAIL : SCE_KERNEL_STOP_SUCCESS;
 }
